+++ packages/core/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
+## [Unreleased]
+
+Updated public api.
+- Exposed enums `LegendEvent` and `Roles`
+- Exposed helper functions from `/packages/core/src/tools.ts`.
+
+# Change Log
 
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.